backend/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
alpaca-py=0.28.0=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.4.0=pypi_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.3.11=haa95532_0
certifi=2024.7.4=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
dnspython=2.6.1=pypi_0
email-validator=2.2.0=pypi_0
expat=2.6.2=hd77b12b_0
fastapi=0.111.0=pypi_0
fastapi-cli=0.0.4=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.5=pypi_0
httptools=0.6.1=pypi_0
httpx=0.27.0=pypi_0
idna=3.7=pypi_0
jinja2=3.1.4=pypi_0
libffi=3.4.4=hd77b12b_1
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
mdurl=0.1.2=pypi_0
msgpack=1.0.8=pypi_0
numpy=2.0.0=pypi_0
openssl=3.0.14=h827c3e9_0
orjson=3.10.6=pypi_0
pandas=2.2.2=pypi_0
pip=24.0=py312haa95532_0
pydantic=2.8.2=pypi_0
pydantic-core=2.20.1=pypi_0
pydantic-settings=2.3.4=pypi_0
pygments=2.18.0=pypi_0
python=3.12.4=h14ffc60_1
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
python-multipart=0.0.9=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
requests=2.32.3=pypi_0
rich=13.7.1=pypi_0
setuptools=69.5.1=py312haa95532_0
shellingham=1.5.4=pypi_0
six=1.16.0=pypi_0
sniffio=1.3.1=pypi_0
sqlite=3.45.3=h2bbff1b_0
sseclient-py=1.8.0=pypi_0
starlette=0.37.2=pypi_0
tk=8.6.14=h0416ee5_0
typer=0.12.3=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.1=pypi_0
ujson=5.10.0=pypi_0
urllib3=2.2.2=pypi_0
uvicorn=0.30.1=pypi_0
vc=14.2=h2eaa2aa_4
vs2015_runtime=14.29.30133=h43f2093_4
watchfiles=0.22.0=pypi_0
websockets=12.0=pypi_0
wheel=0.43.0=py312haa95532_0
xz=5.4.6=h8cc25b3_1
zlib=1.2.13=h8cc25b3_1
